+++ e2e/e2etests/test_sui_token_deposit.go
@@ -20,7 +20,7 @@ func TestSuiTokenDeposit(r *runner.E2ERunner, args []string) {
 	require.NoError(r, err)
 
 	// make the deposit transaction
-	resp := r.SuiDepositFungibleToken(r.EVMAddress(), math.NewUintFromBigInt(amount))
+	resp := r.SuiDepositFungibleToken(r.SuiGateway.PackageID(), r.EVMAddress(), math.NewUintFromBigInt(amount))
 
 	r.Logger.Info("Sui deposit tx: %s", resp.Digest)