+++ zetaclient/chains/sui/signer/tss_owned_object.go
@@ -5,7 +5,10 @@ import (
 	"sync"
 	"time"
 
+	"github.com/block-vision/sui-go-sdk/models"
 	"github.com/pkg/errors"
+
+	"github.com/zeta-chain/node/pkg/contracts/sui"
 )
 
 const tssOwnedObjectTTL = 5 * time.Minute
@@ -35,6 +38,17 @@ func (wc *tssOwnedObject) set(objectID string) {
 
 	wc.objectID = objectID
 	wc.fetchedAt = time.Now()
+}
+
+// withdrawCapID returns the objectID of the WithdrawCap.
+func (s *Signer) withdrawCapID(ctx context.Context) (string, error) {
+	// withdraw cap ID in the chain params is preferred
+	if withdrawCapID := s.gateway.WithdrawCapID(); withdrawCapID != "" {
+		return withdrawCapID, nil
+	}
+
+	// query from Sui network for backward compatibility
+	return s.getWithdrawCapIDCached(ctx)
 }
 
 // getWithdrawCapIDCached getWithdrawCapID with tssOwnedObjectTTL cache.
@@ -70,37 +84,52 @@ func (s *Signer) getWithdrawCapID(ctx context.Context) (string, error) {
 	return objectID, nil
 }
 
-// TODO: https://github.com/zeta-chain/node/issues/4066
-// uncomment below helper functions used for authenticated call
 // getMessageContextIDCached getMessageContextID with tssOwnedObjectTTL cache.
-// func (s *Signer) getMessageContextIDCached(ctx context.Context) (string, error) {
-// 	if s.messageContext.valid() {
-// 		return s.messageContext.objectID, nil
-// 	}
+func (s *Signer) getMessageContextIDCached(ctx context.Context) (string, error) {
+	if s.messageContext.valid() {
+		return s.messageContext.objectID, nil
+	}
 
-//	s.Logger().Std.Info().Msg("messageContext cache expired, fetching new objectID")
+	s.Logger().Std.Info().Msg("MessageContext cache expired, fetching new objectID")
 
-// 	objectID, err := s.getMessageContextID(ctx)
-// 	if err != nil {
-// 		return "", errors.Wrap(err, "unable to get message context ID")
-// 	}
+	objectID, err := s.getMessageContextID(ctx)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to get message context ID")
+	}
 
-// 	s.messageContext.set(objectID)
+	s.messageContext.set(objectID)
 
-//	s.Logger().Std.Info().Str("sui_object_id", objectID).Msg("messageContext objectID fetched")
+	s.Logger().Std.Info().Str("sui_object_id", objectID).Msg("MessageContext objectID fetched")
 
-// 	return objectID, nil
-// }
+	return objectID, nil
+}
 
-// getMessageContextID returns the objectID of the MessageContext. Should belong to TSS address on Sui.
-// func (s *Signer) getMessageContextID(ctx context.Context) (string, error) {
-// 	owner := s.TSS().PubKey().AddressSui()
-// 	structType := s.gateway.MessageContextType()
+// getMessageContextID returns the objectID of the active MessageContext. Should belong to TSS address on Sui.
+func (s *Signer) getMessageContextID(ctx context.Context) (string, error) {
+	nameJSON, err := sui.ActiveMessageContextDynamicFieldName()
+	if err != nil {
+		return "", errors.Wrap(err, "unable to get dynamic field name")
+	}
 
-// 	objectID, err := s.client.GetOwnedObjectID(ctx, owner, structType)
-// 	if err != nil {
-// 		return "", errors.Wrap(err, "unable to get owned object ID")
-// 	}
+	response, err := s.suiClient.SuiXGetDynamicFieldObject(ctx, models.SuiXGetDynamicFieldObjectRequest{
+		ObjectId: s.gateway.ObjectID(),
+		DynamicFieldName: models.DynamicFieldObjectName{
+			Type:  "vector<u8>",
+			Value: nameJSON,
+		},
+	})
+	if err != nil {
+		return "", errors.Wrap(err, "unable to get message context dynamic field object")
+	}
 
-// 	return objectID, nil
-// }
+	if response.Data == nil || response.Data.Content == nil {
+		return "", errors.New("dynamic field object data is nil")
+	}
+
+	messageContextID, err := sui.ParseDynamicFieldValueStr(*response.Data.Content)
+	if err != nil {
+		return "", errors.Wrapf(err, "unable to parse message context ID")
+	}
+
+	return messageContextID, nil
+}